requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
PyMySQL>=1.0.2
SQLAlchemy>=1.4.0
APScheduler>=3.9.0
dingtalk-sdk>=1.0.0
feishu-sdk>=1.0.0
wechatpy>=1.8.0
lxml>=4.9.0
requests-oauthlib>=1.3.0
tweepy>=4.12.0
loguru>=0.6.0
python-dotenv>=0.19.0
PyYAML>=6.0
Flask>=2.0.0
Flask-WTF>=1.0.0
WTForms>=3.0.0